requirements.txt
pandas
nltk
scikit-learn
python-fasthtml